+++ packages/nodes-base/nodes/Linear/shared/GenericFunctions.ts
